+++ Strategies/Sorting/MergeSort.cs
@@ -7,8 +7,8 @@ namespace Projekt1_gr2.Strategies.Sorting;
 public class MergeSort : ISortingStrategy
 {
     public string Name => "MergeSort";
-    private static int comparisons = 0;
-    private static int merges = 0;
+    private static long comparisons = 0;
+    private static long merges = 0;
     public SortStatistics Sort(int[] array) 
     {
         var stats = new SortStatistics{AlgorithmName = Name, Size=array.Length};
